Chevaun Ho Week 4 Assessment 4
Hi Yang,

Sorry for the late submission.

I think I’ve got it. Not sure why I can’t make the other paragraphs
appear in the second blog post. Also not sure if the ‘Read More’ click
and show function is only meant to work for the blog post clicked on,
or for all blog posts.

Thanks,
Chev